Доработал считывание данных из файла

+++ WebUI/Common.Constants/Coefficients.cs
@@ -10,42 +10,42 @@
         /// <summary>
         /// Коэффициент для канала 1
         /// </summary>
-        public const double Ch1 = (9685 - Ch1Low)/_byte;
-        public const double Ch1Low = 8915;
+        public const float Ch1 = (9685f - Ch1Low)/_byte;
+        public const float Ch1Low = 8915f;
 
         /// <summary>
         /// Коэффициент для канала 2
         /// </summary>
-        public const double Ch2 = (9027.85 - Ch2Low)/_byte;
-        public const double Ch2Low = 8073.52;
+        public const float Ch2 = (9027.85f - Ch2Low)/_byte;
+        public const float Ch2Low = 8073.52f;
         /// <summary>
         /// Коэффициент для канала 3
         /// </summary>
-        public const double Ch3 = (8889.66 - Ch3Low)/_byte;
-        public const double Ch3Low = 7394.89;
+        public const float Ch3 = (8889.66f - Ch3Low)/_byte;
+        public const float Ch3Low = 7394.89f;
 
         /// <summary>
         /// Коэффициент для канала 4
         /// </summary>
-        public const double Ch4 = (8246.47 - Ch4Low)/_byte;
-        public const double Ch4Low = 6442.92;
+        public const float Ch4 = (8246.47f - Ch4Low)/_byte;
+        public const float Ch4Low = 6442.92f;
 
         /// <summary>
         /// Коэффициент для канала 5
         /// </summary>
-        public const double Ch5 = (21692.3 - Ch5Low)/_byte;
-        public const double Ch5Low = 11572.7;
+        public const float Ch5 = (21692.3f - Ch5Low)/_byte;
+        public const float Ch5Low = 11572.7f;
 
         /// <summary>
         /// Коэффициент для канала 6
         /// </summary>
-        public const double Ch6 = (13780.8 - Ch6Low)/_byte;
-        public const double Ch6Low = 8170.02;
+        public const float Ch6 = (13780.8f - Ch6Low)/_byte;
+        public const float Ch6Low = 8170.02f;
 
         /// <summary>
         /// Коэффициент для канала 7
         /// </summary>
-        public const double Ch7 = (10046.5 - Ch7Low)/_byte;
-        public const double Ch7Low = 6487.5;
+        public const float Ch7 = (10046.5f - Ch7Low)/_byte;
+        public const float Ch7Low = 6487.5f;
     }
 }
